refactor(*): Tidy up the way collections are accessed
